Fix: probe bfloat16 dtype + use probe_meta_sft path

+++ src/training/grpo.py
@@ -52,7 +52,7 @@ def _load_probe(probe_path: str, hidden_dim: int = 3584, device: str = "cuda"):
     probe = SimpleCorrectnessProbe(hidden_dim=hidden_dim)
     state = torch.load(Path(probe_path) / "best_probe.pt", map_location="cpu", weights_only=True)
     probe.load_state_dict(state)
-    probe.to(device).eval()
+    probe.to(device).to(torch.bfloat16).eval()  # Match model dtype
     print(f"Probe loaded from {probe_path}")
     return probe
